Update testRFI_framework.py
Modified the way of creating the lists so is easier to add new emitters, receivers, etc.
Using the append method of the list.

Moved the generation of sky signal before receive RFI, seems more loical.

Added     for a in antRxL:
                     print(a.Name)
to print the names of the objects generated so you dont have to modify so much code to add many emitters o receivers.